% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utilities.R
\name{check_book_links}
\alias{check_book_links}
\title{Check whether the links in the compiled hanbdooks respond or not}
\usage{
check_book_links(
  book_path,
  no_page_response = "the page you requested could not be found",
  verbose = TRUE
)
}
\arguments{
\item{book_path}{path to the directory with the compiled version of
the hanbdook.}

\item{no_page_response}{a character string to look for in the
server response that a page does not exist. Default is
"the page you requested could not be found".}

\item{verbose}{Should we get progress updates? Default is \code{TRUE}.}
}
\description{
Check whether the links in the compiled hanbdooks respond or not
}
