% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utilities.R
\name{check_module_list}
\alias{check_module_list}
\title{Parity checks for module lists}
\usage{
check_module_list(
  module_list,
  current_session = "21/22",
  modules_with_differences = "EP304",
  no_page_response = "the page you requested could not be found",
  verbose = TRUE,
  check_urls = TRUE
)
}
\arguments{
\item{module_list}{an object of class \code{module_list} as produced by
\code{\link{get_module_list}}.}

\item{current_session}{string characterizing the current academic
session. Default is "21/22".}

\item{modules_with_differences}{module codes to exclude from checks
whether duplicated records have the same infor for "Code",
"Name", "Suspended", "Suspended_Session", "URL", "Notes" and
"Term".}

\item{no_page_response}{a character string to look for in the
server response that a page does not exist. Default is
"the page you requested could not be found".}

\item{verbose}{Should we get progress updates? Default is \code{TRUE}.}
}
\description{
Parity checks for module lists
}
