% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_draft.R
\name{compile_draft}
\alias{compile_draft}
\title{Compile a handbook draft}
\usage{
compile_draft(
  working_dir,
  courses = c("datascience", "morse", "mathstat", "msc"),
  include_source_names = TRUE,
  keep_source = FALSE,
  output_format = "html",
  quiet = TRUE
)
}
\arguments{
\item{working_dir}{the directory where the handbook structure is,
i.e. with directories course/md course/html-* and common/md
common/html-*, as created by the script
download_convert_organize_handbook_files.R.}

\item{courses}{at least one of \code{"datascience"}, \code{"morse"},
\code{"mathstat"}, \code{"msc"} indicating which handbook drafts to
compile. Default is \code{c("datascience", "morse", "mathstat")}.}

\item{include_source_names}{should the draft include notes of what
md file each part of the draft corresponds to? Default is
\code{TRUE}.}

\item{keep_source}{should the draft md files be kept or deleted
after compilation? Default is \code{FALSE}.}

\item{output_format}{either \code{"html"} (default), \code{"word"}, or \code{"pdf"}.}

\item{quiet}{suppress printing of progress? Default is \code{TRUE}.}
}
\description{
Compile a handbook draft
}
\author{
Ioannis Kosmidis \link{aut, cre} \email{ioannis.kosmidis@warwick.ac.uk}
}
