% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursehand.R
\docType{package}
\name{coursehand}
\alias{coursehand}
\title{Methods to maintain the Course Handbooks of the Department of
Statistics at University of Warwick}
\description{
Provides methods and functions to maintain and export the course
handbooks of the Department of Statistics at University of Warwick
using Rmarkdown files. The package depends on a particular
organization of the content, so it will, most probably, not be of
any immediate use to anyone outside the Department of Statistics at
University of Warwick.
}
\author{
Ioannis Kosmidis \email{ioannis.kosmidis@warwick.ac.uk}

Ioannis Kosmidis (aut, cre) \email{ioannis.kosmidis@warwick.ac.uk}
}
