% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utilities.R
\name{get_links_from_html}
\alias{get_links_from_html}
\title{Extract all http and https links form an html file}
\usage{
get_links_from_html(path)
}
\arguments{
\item{path}{the path to the html file. It can be a URL supplied as a string.}
}
\description{
Extract all http and https links form an html file
}
