% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_lists.R
\name{get_module_list}
\alias{get_module_list}
\title{Filter out particular module lists from a master list}
\usage{
get_module_list(
  module_list,
  course = "datascience",
  stream = NA,
  year = 1,
  list = "Core",
  bsc = "Yes",
  mlevel = "Yes"
)
}
\arguments{
\item{module_list}{either the path to \code{module_lists.csv} or a
\code{data.frame}.}

\item{course}{one of "datascience", "morse", "mathstat".}

\item{stream}{either \code{NA} (default) or one of \code{"G30A"}, \code{"G30B"},
\code{"G30C"}, \code{"G30D"}.}

\item{year}{either 1 (default), 2, 3, or 4.}

\item{list}{one of \code{"Core"}, \code{"List A"}, \code{"List B"}, \code{"List C"},
\code{"List D"}, \code{"List E"}, \code{"List F"}, \code{"Optional"}.}

\item{bsc}{either "Yes" (default) or "No", indicating whether the request is about a list that is available for a BSc course or not.}

\item{mlevel}{either "Yes" (default) or "No", indicating whether the request is about a list that is available for an M-level  course or not.}
}
\description{
Filter out particular module lists from a master list
}
