% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_course_Rmd}
\alias{list_course_Rmd}
\title{List Rmd files available for a course}
\usage{
list_course_Rmd(working_dir, course, full_path = TRUE)
}
\arguments{
\item{working_dir}{the path to the working directory.}

\item{course}{one of "datascience", "morse", "mathstat" or "common".}

\item{full_path}{should the full paths to the md files be returned? Default is \code{TRUE}.}
}
\description{
List Rmd files available for a course
}
\author{
Ioannis Kosmidis \link{aut, cre} \email{ioannis.kosmidis@warwick.ac.uk}
}
