% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{Rmd_to_md}
\alias{Rmd_to_md}
\title{Convert Rmd files to md and copy the md contents to the clipboard}
\usage{
Rmd_to_md(Rmd_file, course, working_dir, clipboard = TRUE, quiet = TRUE)
}
\arguments{
\item{Rmd_file}{the name of the Rmd file to be converted to md.}

\item{course}{one of "datascience", "morse", "mathstat" or "common".}

\item{working_dir}{the path to the working directory.}

\item{clipboard}{should the contents of the html file be copied to the clipboard? Default is \code{TRUE}.}

\item{quiet}{suppress printing of progress? Default is \code{TRUE}.}
}
\description{
Convert Rmd files to md and copy the md contents to the clipboard
}
\author{
Ioannis Kosmidis \link{aut, cre} \email{ioannis.kosmidis@warwick.ac.uk}
}
