% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warstguides.R
\docType{package}
\name{warstguides}
\alias{warstguides}
\title{Methods to maintain the Course Handbooks of the Department of
Statistics at University of Warwick}
\description{
Provides methods and functions to maintain and export the course
handbooks of the Department of Statistics at Unviersity of Warwick
using Rmarkdown files. The package depends on a particular
organization of the content, so it wil; most probably not me of any
use to anyone outside the Department of Statistics at Unviersity of
Warwick.
}
\author{
Ioannis Kosmidis \email{ioannis.kosmidis@warwick.ac.uk}
}
