% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_book.R
\name{compile_book}
\alias{compile_book}
\title{Compile a handbook in html and pdf format}
\usage{
compile_book(
  working_dir,
  courses = c("datascience", "morse", "mathstat", "msc"),
  include_source_names = FALSE,
  quiet = TRUE,
  compress = TRUE,
  draft_version = TRUE
)
}
\arguments{
\item{working_dir}{the directory where the handbook structure is,
i.e. with directories course/md course/html-* and common/md
common/html-*, as created by the script
download_convert_organize_handbook_files.R.}

\item{courses}{at least one of \code{"datascience"}, \code{"morse"},
\code{"mathstat"}, \code{"msc"} indicating which handbook drafts to
compile. Default is \code{c("datascience", "morse", "mathstat", "msc")}.}

\item{include_source_names}{should the draft include notes of what
md file each part of the draft corresponds to? Default is
\code{FALSE}.}

\item{quiet}{suppress printing of progress? Default is \code{TRUE}.}

\item{compress}{should the resulting gitbook be compressed. Default is \code{TRUE}.}

\item{draft_version}{should the html version of the handboook have
a "DRAFT" watermark?  Default is \code{TRUE}}
}
\description{
Compile a handbook in html and pdf format
}
\author{
Ioannis Kosmidis (aut, cre) \email{ioannis.kosmidis@warwick.ac.uk}
}
