% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_lists.R
\name{print.module_list}
\alias{print.module_list}
\title{Print a module list as a markdown pipe table}
\usage{
\method{print}{module_list}(x, include = c("Code", "Name", "CATS", "Term"), ...)
}
\arguments{
\item{x}{the output of \code{\link{get_module_list}}}

\item{include}{a subset of \code{"Code"}, \code{"Name"}, \code{"CATS"}, \code{"Term"},
\code{"Req"}, \code{"Other_Streams"}}

\item{...}{currently not used}
}
\description{
Print a module list as a markdown pipe table
}
\author{
Ioannis Kosmidis \link{aut, cre} \email{ioannis.kosmidis@warwick.ac.uk}
}
