% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_lists.R
\name{summary.module_list}
\alias{summary.module_list}
\title{Various modes of analyses for module lists}
\usage{
\method{summary}{module_list}(x, type = "module_code_xtabs", ...)
}
\arguments{
\item{x}{an object of class \code{module_list} as produced by
\code{\link{get_module_list}}.}

\item{type}{\code{"module_code_xtabs"} for cross-tabulation of
department and level of the modules in \code{x} (default), or
\code{"word_cloud"} for a world cloud of the unique module names in
\code{x}.}

\item{...}{if \code{type = "word_cloud"} further arguments to be passed
to \code{wordcloud::wordcloud()}}
}
\value{
If \code{type = "module_code_xtabs"}, a cross-tabulation of departments
and module levels as according to the module coding used at
University of Warwick. See
\url{https://warwick.ac.uk/fac/sci/statistics/currentstudents/handbooks/morse/modulecodes/}
for details.

If \verb{type = "word_cloud", then a world cloud of the unique module names in }x` is produced.
}
\description{
Various modes of analyses for module lists
}
\author{
Ioannis Kosmidis \link{aut, cre} \email{ioannis.kosmidis@warwick.ac.uk}
}
